+++ app/hermes/views.py
@@ -2,7 +2,6 @@ from fastapi import APIRouter, Header
 from fastapi.exceptions import HTTPException, RequestValidationError
 from starlette.requests import Request
 
-from app.hermes.tasks import get_next_sales_person
 from app.models import Admin, Company
 
 main_router = APIRouter()
@@ -62,6 +61,8 @@ EU_COUNTRIES = [
     'JE',
     'AX',
 ]
+
+
 # EU Countries:
 # MD: Moldova, BG: Bulgaria, DE: Germany, AL: Albania, ME: Montenegro, ES: Spain, SE: Sweden, AD: Andorra, MT: Malta,
 # CZ: Czech Republic, GB: United Kingdom, GI: Gibraltar, CY: Cyprus, MC: Monaco, RU: Russia, IE: Ireland, FR: France,
@@ -70,6 +71,23 @@ EU_COUNTRIES = [
 # FO: Faroe Islands, SI: Slovenia, UA: Ukraine, FI: Finland, AT: Austria, BA: Bosnia and Herzegovina, GR: Greece,
 # GG: Guernsey, EE: Estonia, SM: San Marino, VA: Vatican City, IT: Italy, SK: Slovakia, LT: Lithuania,
 # IM: Isle of Man, NO: Norway, LV: Latvia, RO: Romania, SJ: Svalbard and Jan Mayen, JE: Jersey, AX: Åland Islands
+
+
+async def get_next_sales_person(admins, latest_sales_person_id):
+    """
+    @param admins: a list of Admin objects
+    @param latest_sales_person_id: the ID of the latest sales Admin object
+    @return: the ID of the next sales Admin object
+    """
+    admin_ids = [a.id async for a in admins.filter(is_sales_person=True).order_by('id')]
+    if latest_sales_person_id:
+        try:
+            next_person_id = admin_ids[admin_ids.index(latest_sales_person_id) + 1]
+        except (IndexError, ValueError):
+            next_person_id = admin_ids[0]
+    else:
+        next_person_id = admin_ids[0]
+    return next_person_id
 
 
 @main_router.get('/choose-roundrobin/sales/', name='Decide which sales person to assign to a new signup')
